[update] 敵がタワーに侵入してきたら、タワーのライフをー１するクラスを作成した。 [update] タワーのコンポーネントがシーン上にある時、どこからでもアクセスできるようにした。 [update] TowerのコンポーネントやPrefabをBattleフォルダの下に移動させた。
