--- Assets/Tower.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b4036f20ab612574cae086cabb04c21e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
